+++ src/app/cli/src/cli_commands.rs
@@ -527,7 +527,7 @@ pub fn command_needs_transaction(args: &cli::Cli) -> bool {
             cli::SystemSubCommand::GenerateToken(_) => true,
             _ => false,
         },
-        cli::Command::Add(_) | cli::Command::Delete(_) => true,
+        cli::Command::Add(_) | cli::Command::Delete(_) | cli::Command::Rename(_) => true,
         _ => false,
     }
 }