+++ CHANGELOG.md
@@ -18,6 +18,7 @@ Recommendation: for ease of reading, use the following order:
 - `kamu repo alias list`: added JSON output alongside with other formats mentioned in the command's help
 - Private Datasets, `DatasetEntry` integration that will allow us to build dataset indexing
   - Extend `DatasetLifecycleMessageCreated` with `dataset_name` field
+  - Introducing `DatasetLifecycleMessageRenamed`
 
 ## [0.204.4] - 2024-09-30
 ### Changed